feat: 파일 확장자에 따라 뷰 네임 처리하는 ViewProcessor 클래스 생성
