+++ src/main/java/webserver/viewProcessor/ViewProcessor.java
@@ -1,0 +1,9 @@
+package webserver.viewProcessor;
+
+import webserver.httpResponse.HttpResponse;
+
+public interface ViewProcessor {
+    boolean isSupported(String viewName);
+
+    String process(String viewName, HttpResponse httpResponse);
+}
